Exception handling

+++ src/main/java/br/com/luna/services/ProductService.java
@@ -1,14 +1,19 @@
 package br.com.luna.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import br.com.luna.dto.ProductDTO;
 import br.com.luna.entities.Product;
 import br.com.luna.repositories.ProductRepository;
+import br.com.luna.services.exceptions.DatabaseException;
+import br.com.luna.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,8 +23,9 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {	
-		Product product = repository.findById(id).get();
-		return new ProductDTO(product);
+			Product product = repository.findById(id).orElseThrow(
+					() -> new ResourceNotFoundException("Recurso não encontrado"));
+			return new ProductDTO(product);		
 		
 	}
 	
@@ -42,16 +48,29 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO update(Long id, ProductDTO dto) {		
+		try {
 		Product entity = repository.getReferenceById(id);
 		copyDtoToEntity(dto, entity);
 		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		return new ProductDTO(entity);		
+		}
+		catch(EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 		
 	}
 	
-	@Transactional
-	public void delete(Long id){	
-		repository.deleteById(id);		
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id){
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {		
+		repository.deleteById(id);			
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
